refactor : User, Owner, Room 엔티티들의 경로 수정
전반적인 패키지 구조 통일을 하여 기존 엔티티 클래스들의 경로를 domain.entity 로 수정
