+++ studycafe/src/main/java/com/scascanner/studycafe/domain/entity/Owner.java
@@ -1,4 +1,4 @@
-package com.scascanner.studycafe.domain;
+package com.scascanner.studycafe.domain.entity;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -8,19 +8,26 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import java.time.LocalDate;
-import java.util.Date;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User {
+public class Owner {
 
-    @Id @GeneratedValue
-    @Column(name = "user_id")
+    @Id
+    @GeneratedValue
+    @Column(name = "owner_id")
     private Long id;
 
     private String email;
+
     private String password;
+
     private String nickname;
+
     private String name;
     private LocalDate birthday;
+
+    public Long getId() {
+        return id;
+    }
 }
